- DOR-428 Updated warning level on MSVC to W4.

+++ dorado/cli/cli_utils.h
@@ -224,8 +224,8 @@ inline void process_internal_arguments(const ArgParser& parser) {
 template <class Options>
 Options process_minimap2_arguments(const ArgParser& parser, const Options& dflt) {
     Options res;
-    res.kmer_size = parser.visible.get<int>("k");
-    res.window_size = parser.visible.get<int>("w");
+    res.kmer_size = short(parser.visible.get<int>("k"));
+    res.window_size = short(parser.visible.get<int>("w"));
     res.index_batch_size = cli::parse_string_to_size(parser.visible.get<std::string>("I"));
     res.print_secondary = cli::parse_yes_or_no(parser.visible.get<std::string>("secondary"));
     res.best_n_secondary = parser.visible.get<int>("N");
